Starting refactor with login page.
Backing up old main.dart for reference.
To delete later.
